Slit `stringFrom` and `stringTo` unit tests
