+++ test/test.vcxproj
@@ -165,6 +165,8 @@
     <ClCompile Include="Dictionary.test.cpp" />
     <ClCompile Include="Filesystem.test.cpp" />
     <ClCompile Include="ParserHelper.test.cpp" />
+    <ClCompile Include="StringFrom.test.cpp" />
+    <ClCompile Include="StringTo.test.cpp" />
     <ClCompile Include="StringUtils.test.cpp" />
     <ClCompile Include="StringValue.test.cpp" />
     <ClCompile Include="Utility.test.cpp" />
